MOved authorization logic to new project
