+++ BobsBookStore.sln
@@ -24,6 +24,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BookInventory.ImageWorkflow
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BookInventory.Repository.IntegrationTests", "tests\BookInventory\BookInventory.Repository.IntegrationTests\BookInventory.Repository.IntegrationTests.csproj", "{68E54916-1D7F-467F-ADE5-C3355BEEB9E4}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BookInventory.Authorization", "src\BookInventory\BookInventory.Authorization\BookInventory.Authorization.csproj", "{89EB24BC-2560-47EB-9D35-C65C0B267832}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -94,5 +96,9 @@ Global
 		{68E54916-1D7F-467F-ADE5-C3355BEEB9E4}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{68E54916-1D7F-467F-ADE5-C3355BEEB9E4}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{68E54916-1D7F-467F-ADE5-C3355BEEB9E4}.Release|Any CPU.Build.0 = Release|Any CPU
+		{89EB24BC-2560-47EB-9D35-C65C0B267832}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{89EB24BC-2560-47EB-9D35-C65C0B267832}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{89EB24BC-2560-47EB-9D35-C65C0B267832}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{89EB24BC-2560-47EB-9D35-C65C0B267832}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
