+++ src/BookInventory/BookInventory.Api/BookInventory.Api.csproj
@@ -20,13 +20,9 @@
         <PackageReference Include="AWS.Lambda.Powertools.Logging" Version="1.4.4" />
         <PackageReference Include="AWS.Lambda.Powertools.Tracing" Version="1.3.2" />
 		<PackageReference Include="AWS.Lambda.Powertools.Metrics" Version="1.5.3" />
-		<PackageReference Include="AWSSDK.CognitoIdentityProvider" Version="3.7.303.13" />
 		<PackageReference Include="AWSSDK.Rekognition" Version="3.7.301.27" />
 		<PackageReference Include="AWSSDK.S3" Version="3.7.307.25" />
         <PackageReference Include="FluentValidation.DependencyInjectionExtensions" Version="11.8.1" />
-        <PackageReference Include="JsonWebToken" Version="1.9.4" />
-        <PackageReference Include="Microsoft.IdentityModel.Protocols.OpenIdConnect" Version="7.5.1" />
-        <PackageReference Include="System.IdentityModel.Tokens.Jwt" Version="7.5.1" />
     </ItemGroup>
     <ItemGroup>
       <ProjectReference Include="..\..\..\cdk\src\SharedConstructs\SharedConstructs.csproj" />
