+++ src/BookInventory/BookInventory.Api/serverless.template
@@ -147,23 +147,6 @@
           }
         }
       }
-    },
-    "BookInventoryApiFunctionsBookInventoryAuthorizerGenerated": {
-      "Type": "AWS::Serverless::Function",
-      "Metadata": {
-        "Tool": "Amazon.Lambda.Annotations"
-      },
-      "Properties": {
-        "Runtime": "dotnet6",
-        "CodeUri": ".",
-        "MemorySize": 256,
-        "Timeout": 30,
-        "Policies": [
-          "AWSLambdaBasicExecutionRole"
-        ],
-        "PackageType": "Zip",
-        "Handler": "BookInventory.Api::BookInventory.Api.Functions_BookInventoryAuthorizer_Generated::BookInventoryAuthorizer"
-      }
     }
   }
 }
