+++ src/BookInventory/BookInventory.Authorization/BookInventory.Authorization.csproj
@@ -1,0 +1,39 @@
+<Project Sdk="Microsoft.NET.Sdk">
+    <PropertyGroup>
+        <OutputType>Library</OutputType>
+        <TargetFramework>net8.0</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+        <GenerateRuntimeConfigurationFiles>true</GenerateRuntimeConfigurationFiles>
+        <AWSProjectType>Lambda</AWSProjectType>
+        <!-- This property makes the build directory similar to a publish directory and helps the AWS .NET Lambda Mock Test Tool find project dependencies. -->
+        <CopyLocalLockFileAssemblies>true</CopyLocalLockFileAssemblies>
+        <!-- Generate ready to run images during publishing to improvement cold starts. -->
+        <PublishReadyToRun>true</PublishReadyToRun>
+    </PropertyGroup>
+    <ItemGroup>
+        <PackageReference Include="Amazon.Lambda.Core" Version="2.2.0"/>
+        <PackageReference Include="Amazon.Lambda.APIGatewayEvents" Version="2.7.0"/>
+        <PackageReference Include="Amazon.Lambda.Serialization.SystemTextJson" Version="2.4.1"/>
+        <PackageReference Include="Amazon.Lambda.Annotations" Version="1.3.0"/>
+        <PackageReference Include="AWS.Lambda.Powertools.Logging" Version="1.4.4" />
+        <PackageReference Include="AWS.Lambda.Powertools.Tracing" Version="1.3.2" />
+        <PackageReference Include="AWS.Lambda.Powertools.Metrics" Version="1.5.3" />
+        <PackageReference Include="AWSSDK.CognitoIdentityProvider" Version="3.7.303.13" />
+        <PackageReference Include="JsonWebToken" Version="1.9.4" />
+        <PackageReference Include="Microsoft.IdentityModel.Protocols.OpenIdConnect" Version="7.5.1" />
+        <PackageReference Include="System.IdentityModel.Tokens.Jwt" Version="7.5.1" />
+        <PackageReference Include="AWSSDK.Extensions.NETCore.Setup" Version="3.7.300" />
+    </ItemGroup>
+    <!-- 
+      The FrameworkReference is used to reduce the deployment bundle size by not having to include 
+      dependencies like Microsoft.Extensions.DependencyInjection. The Microsoft.AspNetCore.App
+      which is available in the Managed .NET Lambda runtime already includes those assemblies.
+    -->
+    <ItemGroup>
+        <FrameworkReference Include="Microsoft.AspNetCore.App"/>
+    </ItemGroup>
+    <ItemGroup>
+      <ProjectReference Include="..\..\..\cdk\src\SharedConstructs\SharedConstructs.csproj" />
+    </ItemGroup>    
+</Project>
